image moved to assets folder
