+++ src/Common/Pager/index.js
@@ -4,8 +4,8 @@ import { setStatus } from "../../MoviesSlice";
 import useQueryParameter from "../../useQueryParameter";
 import useReplaceQueryParameter from "../../useReplaceQueryParameter";
 import { pageKey } from "../../functionsAndKeys/keys";
-import arrow from "./arrow.svg";
-import disabledArrow from "./disabledArrow.svg";
+import arrow from "../../assets/arrow.svg";
+import disabledArrow from "../../assets//disabledArrow.svg";
 import {
     Wrapper,
     ButtonContainer,
